prefer values from old tree when merging, COUCHDB-968
This commit represents a substantial refactor of the key tree merging
logic, some of which is not strictly necessary for the resolution of
COUCHDB-968.

Two etap test cases checking the ability to merge in a non-linear tree
are removed because the functionality is no longer supported.  CouchDB
only ever merged a linear revision history into an existing revision
tree.