usort the infos during compaction to remove dupes, COUCHDB-968
This is not a bulletproof solution; it only removes dupes when the
they appear in the same batch of 1000 updates.  However, for dupes
that show up in _all_docs the probability of that happening is quite
high.  If the dupes are only in _changes a user may need to compact
twice, once to get the dupes ordered together and a second time to
remove them.

A more complete solution would be to trigger the compaction in "retry"
mode, but this is siginificantly slower.