Ensure dreyfus index module is loaded
In development mode (dev/run) the dreyfus_index module is loaded by
startup. For release mode no such loading is done, instead modules are
loaded "on demand". In mango code, all use of dreyfus_index is protected
by checking if module is loaded (by calling module_loaded(dreyfus_index)).

This commit adds startup.erl containing code to be run during startup
of couchdb release (rel/couchdb/bin/couchdb) that fixes the error
occuring from dreyfus_index not being loaded.

For a seasoned couchdb or erlang developer this might seem to be a
trivial thing to fix, as for them the use of .erlang in working directory
is a well-known feature. Others might find this useful as it makes
couchdb with dreyfus working "out-of-the-box" in release mode too.